% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IALS_data.R
\docType{data}
\name{IALS_data}
\alias{IALS_data}
\title{International Adult Literacy Survey (IALS) for 13 countries}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{Prose}{On prose scale, the percentage of individuals who could not reach a basic level of literacy in each country.}
 \item{Document}{On document scale, the percentage of individuals who could not reach a basic level of literacy in each country.}
 \item{Quantitative}{On quantitative scale, the percentage of individuals who could not reach a basic level of literacy in each country.}
 \item{Country}{Specify the country}
 \item{Gender}{Specify the gender}
}
}
\usage{
data(IALS_data)
}
\description{
The data is obtained from the International Adult Literacy Survey (IALS),
collected in 13 countries on Prose, Document, and Quantitative scales between 1994 and 1995.
The data are reported as the percentage of individuals who could not reach a basic level of literacy in each country.
}
\examples{

data(IALS_data)
head(IALS_data)
}
\references{
Sofroniou, N., Hoad, D., & Einbeck, J. (2008).  League tables for literacy survey data based on random effect models.
            In: Proceedings of the 23rd International Workshop on Statistical Modelling, Utrecht; pp. 402-405.
}
\keyword{datasets}
