% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_data.R
\docType{data}
\name{trading_data}
\alias{trading_data}
\title{A set of import and export data in 44 countries.}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{import}{The fetus from the same twins share the same id number.}
 \item{export}{frequency of self-touch for each fetus.}
 \item{country}{frequency of twin-to-twin for each fetus.}
}
}
\source{
Trade in Goods and Services. \url{https://data.oecd.org/trade/trade-in-goods-and-services.htm}. Accessed on 2023-05-29.
}
\usage{
data(trading_data)
}
\description{
The variables are given as the percentage of imports and exports in relation to the overall GDP.
The data set comprises data from 44 countries (for our analysis),
we specifically selected the time period between 2018 and 2022.
}
\examples{

data(trading_data)
head(trading_data)
}
\keyword{datasets}
